Implemented functional button images
I created a new class, PressedButton, which renders the selected UV when the button is held, and renders nothing (ie, the background) when it is released.

The UI is inconsistent since the rest of the UI shows highlights around the buttons, but I think that is out of scope for this change.