Fix Crafting
This fixes the crafting that was broken in the porting process.

In 1.19.1, #minecraft:stripped_log was introduced, but in 1.19.3 it was then removed... and never readded. So to reference stripped logs as a crafting ingredient, it looks like NeoForge adds some kind of #c:stripped_logs in its place. So I used that. It might not work; it's a nonsense recipe anyway so I'll probably change it at a later date.