Fixes data generation for mod
This version differs from a previous PR; upstream fixed the underlying bug causing Gradle to malfunction, meaning that I was able to simply update the dependency versions. This should have no bearing on mod compatibility with Simple Voice Chat versions other than the most recent version as it only affects the build of datagen in a development environment.

* Fixes the data generation by using updated SimpleVoiceChat maven resources
* Fixes the blockstates that are missing that were manually added to generated files
* Fixes the translations that are missing that were manually added to generated files