PTT Handheld Fix
I made quite a few changes and refactors with this one. The major change is that we no longer use the client tick handler to run click events and instead rely on the use event to trigger PTT. This comes with the side effect of allowing PTT to be handled with a release timer (which is realistic radio operation).

Testing concluded that both a controller and mouse worked for this.

I also added a line in the config to control the hold time after PTT is released.